+++ fabric/src/main/java/io/github/sakurawald/core/gui/inspection/ModulesInspectionGui.java
@@ -1,4 +1,4 @@
-package io.github.sakurawald.core.gui;
+package io.github.sakurawald.core.gui.inspection;
 
 
 import eu.pb4.sgui.api.elements.GuiElementBuilder;
@@ -6,6 +6,8 @@ import eu.pb4.sgui.api.elements.GuiElementInterface;
 import eu.pb4.sgui.api.gui.SimpleGui;
 import io.github.sakurawald.core.auxiliary.ReflectionUtil;
 import io.github.sakurawald.core.auxiliary.minecraft.TextHelper;
+import io.github.sakurawald.core.command.structure.CommandDescriptor;
+import io.github.sakurawald.core.gui.PagedGui;
 import io.github.sakurawald.core.manager.impl.module.ModuleManager;
 import io.github.sakurawald.core.structure.Pair;
 import io.github.sakurawald.module.initializer.ModuleInitializer;
@@ -46,11 +48,11 @@ public class ModulesInspectionGui extends PagedGui<Pair<String, Boolean>> {
     protected GuiElementInterface toGuiElement(Pair<String, Boolean> entity) {
         List<Text> lore = new ArrayList<>();
 
-        /* Add module enable status. */
+        /* Attach module enable status. */
         boolean moduleEnable = entity.getValue();
         lore.add(TextHelper.getTextByKey(getPlayer(), "module.enable.status", moduleEnable));
 
-        /* Extract the @Document from class annotation above module initializer. */
+        /* Attach @Document information above module initializer. */
         String modulePathString = entity.getKey();
         Class<? extends ModuleInitializer> moduleInitializerClass = ModuleManager.MODULE_INITIALIZER_CLASS_BY_MODULE_PATH_STRING.getOrDefault(modulePathString, null);
         if (moduleInitializerClass != null) {
@@ -61,13 +63,40 @@ public class ModulesInspectionGui extends PagedGui<Pair<String, Boolean>> {
             }
         }
 
+        /* Attach registered commands info. */
+        PagedGui<CommandDescriptor> commandsRegisteredByThisModuleGUI = attachModuleCommands(modulePathString, lore);
+
         Item itemMaterial = moduleEnable ? Items.GREEN_STAINED_GLASS : Items.RED_STAINED_GLASS;
         Text itemName = Text.literal(modulePathString).formatted(Formatting.YELLOW);
         return new GuiElementBuilder()
             .setItem(itemMaterial)
             .setName(itemName)
             .setLore(lore)
+            .setCallback(() -> onClickCommandDescriptor(modulePathString, commandsRegisteredByThisModuleGUI))
             .build();
+    }
+
+    private void onClickCommandDescriptor(String modulePathString, PagedGui<CommandDescriptor> commandsRegisteredByThisModuleGUI) {
+
+       /* Click to open the registered commands. */
+       commandsRegisteredByThisModuleGUI.open();
+
+    }
+
+    private PagedGui<CommandDescriptor> attachModuleCommands(String modulePathString, List<Text> lore) {
+        PagedGui<CommandDescriptor> commandsRegisteredByThisModuleGUI = CommandDescriptorGui
+            .makeDefault(getGui(), getPlayer())
+            .search(it -> it.getSourceModulePath().equals(modulePathString));
+
+
+        lore.add(TextHelper.TEXT_EMPTY);
+        int registeredCommandsCount = commandsRegisteredByThisModuleGUI.getEntities().size();
+        lore.add(TextHelper.getTextByKey(getPlayer(), "module.registered_commands", registeredCommandsCount));
+        if (registeredCommandsCount != 0) {
+            lore.add(TextHelper.getTextByKey(getPlayer(), "prompt.click.see_it.left_click"));
+        }
+
+        return commandsRegisteredByThisModuleGUI;
     }
 
     private static @Nullable String getDocumentString(Class<? extends ModuleInitializer> moduleInitializerClass) {