+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 - [fuji] feature: new `/fuji inspect permissions-and-metas` command. to inspect all `permissions` and `metas`.
 - [fuji] feature: add `document string` for all configuration models. (Issue `/fuji inspect configurations` command to see it.)
 - [fuji] feature: add `document string` for all fuji commands. (Issue `/fuji inspect fuji-commands` to see it.)
+- [fuji] feature: now will display `Commands registered by this module` information in `/fuji inspect modules` command.
 - [fuji] feature: display `possible values` information for `enum type` in inspection GUI.
-- [core] feature: unify the document style for `/fuji inspect fuji-commands` command.
-- [core] feature: unify the document style for `/fuji inspect modules` command.
-- [core] feature: unify the document style for `/fuji inspect configurations` command.
+- [fuji] feature: unify the document style for `/fuji inspect fuji-commands` command.
+- [fuji] feature: unify the document style for `/fuji inspect modules` command.
+- [fuji] feature: unify the document style for `/fuji inspect configurations` command.
 - [fuji] fix: escape the style tags for `object.name`. (For `/fuji inspect configurations` command)
 - [fuji] feature: use different item for `Enum` type. (For `/fuji inspect configurations` command)